Crate rustc_back

Unstable (rustc_private)

Some stuff used by rustc that doesn't have many dependencies

Originally extracted from rustc::back, which was nominally the compiler 'backend', though LLVM is rustc's backend, so rustc_back is really just odds-and-ends relating to code gen and linking. This crate mostly exists to make rustc smaller, so we might put more 'stuff' here in the future. It does not have a dependency on rustc_llvm.

FIXME: Split this into two crates: one that has deps on syntax, and one that doesn't; the one that doesn't might get decent parallel build speedups.

Modules

dynamic_lib 	 [Unstable]
Dynamic library facilities.

rpath 	 [Unstable]

sha2 	 [Unstable]
This module implements only the Sha256 function since that is all that is needed for internal use. This implementation is not intended for external use or for any use where security is important.

slice 	 [Unstable]

target 	 [Unstable]
Flexible target specification.

tempdir 	 [Unstable]