Struct rustc_back::dynamic_lib::DynamicLibrary

pub struct DynamicLibrary { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl DynamicLibrary

pub fn open(filename: Option<&Path>) -> Result<DynamicLibrary, String>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Lazily open a dynamic library. When passed None it gives a handle to the calling process

pub fn envvar() -> &'static str

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns the environment variable for this process's dynamic library search path

pub unsafe fn symbol<T>(&self, symbol: &str) -> Result<*mut T, String>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Accesses the value at the symbol of the dynamic library.