Module rustc_back::sha2

Unstable (rustc_private)

This module implements only the Sha256 function since that is all that is needed for internal use. This implementation is not intended for external use or for any use where security is important.

Structs

Sha256 	 [Unstable]
The SHA-256 hash algorithm

Traits

Digest 	 [Unstable]
The Digest trait specifies an interface common to digest functions, such as SHA-1 and the SHA-2 family of digest functions.