Struct rustc_back::tempdir::TempDir

pub struct TempDir { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A wrapper for a path to temporary directory implementing automatic scope-based deletion.

Methods

impl TempDir

pub fn new_in<P: AsRef<Path>>(tmpdir: P, prefix: &str) -> Result<TempDir>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Attempts to make a temporary directory inside of tmpdir whose name will have the prefix prefix. The directory will be automatically deleted once the returned wrapper is destroyed.

If no directory can be created, Err is returned.

pub fn new(prefix: &str) -> Result<TempDir>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Attempts to make a temporary directory inside of env::temp_dir() whose name will have the prefix prefix. The directory will be automatically deleted once the returned wrapper is destroyed.

If no directory can be created, Err is returned.

pub fn into_path(self) -> PathBuf

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Unwrap the wrapped std::path::Path from the TempDir wrapper. This discards the wrapper so that the automatic deletion of the temporary directory is prevented.

pub fn path(&self) -> &Path

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Access the wrapped std::path::Path to the temporary directory.