Struct rustc_back::dynamic_lib::DynamicLibrary

pub struct DynamicLibrary {
    // some fields omitted
}

Unstable (rustc_private)

Methods

impl DynamicLibrary

fn open(filename: Option<&Path>) -> Result<DynamicLibrary, String>

Unstable (rustc_private)

Lazily open a dynamic library. When passed None it gives a handle to the calling process

fn prepend_search_path(path: &Path)

Unstable (rustc_private)

Prepends a path to this process's search path for dynamic libraries

fn create_path(path: &[PathBuf]) -> OsString

Unstable (rustc_private)

From a slice of paths, create a new vector which is suitable to be an environment variable for this platforms dylib search path.

fn envvar() -> &'static str

Unstable (rustc_private)

Returns the environment variable for this process's dynamic library search path

fn search_path() -> Vec<PathBuf>

Unstable (rustc_private)

Returns the current search path for dynamic libraries being used by this process

unsafe fn symbol<T>(&self, symbol: &str) -> Result<*mut T, String>

Unstable (rustc_private)

Accesses the value at the symbol of the dynamic library.