Struct rustc_back::target::Target

pub struct Target {
    pub llvm_target: String,
    pub target_endian: String,
    pub target_pointer_width: String,
    pub target_os: String,
    pub target_env: String,
    pub target_vendor: String,
    pub arch: String,
    pub data_layout: String,
    pub options: TargetOptions,
}

Unstable (rustc_private)

Everything rustc knows about how to compile for a specific target.

Every field here must be specified, and has no default value.

Fields
llvm_target: String
Unstable (rustc_private)

Target triple to pass to LLVM.

target_endian: String
Unstable (rustc_private)

String to use as the target_endian cfg variable.

target_pointer_width: String
Unstable (rustc_private)

String to use as the target_pointer_width cfg variable.

target_os: String
Unstable (rustc_private)

OS name to use for conditional compilation.

target_env: String
Unstable (rustc_private)

Environment name to use for conditional compilation.

target_vendor: String
Unstable (rustc_private)

Vendor name to use for conditional compilation.

arch: String
Unstable (rustc_private)

Architecture to use for ABI considerations. Valid options: "x86", "x86_64", "arm", "aarch64", "mips", "powerpc", and "powerpc64".

data_layout: String
Unstable (rustc_private)

Data layout to pass to LLVM.

options: TargetOptions
Unstable (rustc_private)

Optional settings with defaults.

Methods

impl Target

fn adjust_abi(&self, abi: Abi) -> Abi

Unstable (rustc_private)

Given a function ABI, turn "System" into the correct ABI for this target.

fn from_json(obj: Json) -> Target

Unstable (rustc_private)

Load a target descriptor from a JSON object.

fn search(target: &str) -> Result<Target, String>

Unstable (rustc_private)

Search RUST_TARGET_PATH for a JSON file specifying the given target triple. Note that it could also just be a bare filename already, so also check for that. If one of the hardcoded targets we know about, just return it directly.

The error string could come from any of the APIs called, including filesystem access and JSON decoding.